// src/app/class-type.ts
export class Select {
    category : string;
    subCategory : string;
    tm : string;
};

export class LinkArray{
    link: string;
    value: string;
};

export class CategorySelectors {
    value: string;
    display: string;
};

export class SubCategorySelectors {
    value: string;
    display: string;
};


export class ArrayTM  {
    value: string;
    display: string;
};

export class InstrumentSelector{
    value: string;
    display: string;
    picture: string;
    text: string;
};